launchers: Avoid MockVault class duplications
