+++ samples/other/streaming/src/test/java/org/eclipse/dataspaceconnector/transfer/demo/protocols/fixture/AbstractDemoTransferTest.java
@@ -15,16 +15,12 @@
 package org.eclipse.dataspaceconnector.transfer.demo.protocols.fixture;
 
 import org.eclipse.dataspaceconnector.junit.launcher.EdcExtension;
+import org.eclipse.dataspaceconnector.junit.launcher.MockVault;
 import org.eclipse.dataspaceconnector.spi.security.PrivateKeyResolver;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
-import org.eclipse.dataspaceconnector.spi.security.VaultResponse;
 import org.eclipse.dataspaceconnector.spi.transfer.TransferWaitStrategy;
-import org.jetbrains.annotations.Nullable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
-
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 
 import static org.easymock.EasyMock.niceMock;
 
@@ -49,27 +45,6 @@ public abstract class AbstractDemoTransferTest {
 
         // register a wait strategy of 1ms to speed up the interval between transfer manager iterations
         extension.registerServiceMock(TransferWaitStrategy.class, () -> 1);
-    }
-
-    protected static class MockVault implements Vault {
-        private final Map<String, String> secrets = new ConcurrentHashMap<>();
-
-        @Override
-        public @Nullable String resolveSecret(String key) {
-            return secrets.get(key);
-        }
-
-        @Override
-        public VaultResponse storeSecret(String key, String value) {
-            secrets.put(key, value);
-            return VaultResponse.OK;
-        }
-
-        @Override
-        public VaultResponse deleteSecret(String key) {
-            secrets.remove(key);
-            return VaultResponse.OK;
-        }
     }
 
 }
